Redesigned folders, worked on automatic delete events, logic works but wrong implementation. Need to store it in app.tsx page instead of admin
