+++ .env.template
@@ -1,4 +1,3 @@
 # .env.example
-DATABASE_URL=your_database_url_here
-API_KEY=your_api_key_here
-SECRET_KEY=your_secret_key_here
+VITE_SUPABASE_URL=your_database_url_here
+VITE_SUPABASE_KEY=your_api_key_here
