+++ src/components/Admin/AdminJobs.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 
 type Job ={
