+++ src/components/Admin/AdminNews.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 
 type News = {
