+++ src/components/Admin/SupportManager.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 
 type SupportMessage = {
     id: number;
