+++ src/main.tsx
@@ -1,16 +1,16 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App.tsx";
+import App from "./Pages/App.tsx";
 import { Route, Routes, Navigate, HashRouter } from "react-router-dom";
 import Layout from "./components/layout.tsx";
-import FAQ from "./FAQ.tsx";
-import ForBedrifter from "./ForCompanies.tsx";
-import Jobbtorget from "./Jobs.tsx";
-import OmOss from "./AboutUs.tsx";
-import ContactUs from "./ContactUs.tsx";
-import Admin from "./admin.tsx";
-import NotFound from "./NotFound.tsx";
+import FAQ from "./Pages/FAQ.tsx";
+import ForBedrifter from "./Pages/ForCompanies.tsx";
+import Jobbtorget from "./Pages/Jobs.tsx";
+import OmOss from "./Pages/AboutUs.tsx";
+import ContactUs from "./Pages/ContactUs.tsx";
+import Admin from "./Pages/admin.tsx";
+import NotFound from "./Pages/NotFound.tsx";
 import { AuthProvider, useAuth } from "./auth.tsx"; // ✅ Import authentication
 
 // ✅ Private Route Component
